requirements.txt
sentence-transformers==2.5.1
langchain
flask
pypdf
python-dotenv
pinecone-client>=3.0.0,<4.0.0
langchain-pinecone
langchain_community
langchain_openai
langchain-huggingface
openai>=1.0.0
cryptography
gunicorn